Added check so tips are not put in clusters of 1
When bootstrap threshold set to 0, single sequences can be put into clusters of 1, as tips are considered identically to internal nodes. Now, there is a check so that if the node being compared is a tip, it is never put in a cluster of only itself.  (This by definition is not a cluster)