Fixed CSV bug, added outfile option
Fixed a bug where CSV annotation info files would annotate instead of overwrite when new parameters selected. 
Also, added a new feature that can be turned on in Wind.java that allows all selected clusters to also be printed to one FigTree file to allow them to be easier to browse. Only available when only 1 data file is loaded.